Dev_27_11_2018_1

+++ app/Http/Controllers/API/FixedAssetDepreciationMasterAPIController.php
@@ -410,6 +410,10 @@ class FixedAssetDepreciationMasterAPIController extends AppBaseController
             return $this->sendError('Fixed Asset Depreciation Master not found');
         }
 
+        if($fixedAssetDepreciationMaster->isDepProcessingYN == 0){
+            return $this->sendError('Depreciation is still running', 500);
+        }
+
         $fixedAssetDepreciationMaster->delete();
 
         return $this->sendResponse($id, 'Fixed Asset Depreciation Master deleted successfully');
